Move Notification logic to core and improve notification of authentication message logic
