+++ xcape-api/src/main/java/com/chadev/xcape/api/controller/request/AuthenticationRequest.java
@@ -1,5 +1,6 @@
 package com.chadev.xcape.api.controller.request;
 
+import com.chadev.xcape.core.service.notification.NotificationTemplateEnum;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,4 +15,7 @@ public class AuthenticationRequest {
     private String reservationId;
     private String recipientNo;
 
+    public NotificationTemplateEnum.AuthenticationParam getAuthenticationParam(String authenticationNumber) {
+        return new NotificationTemplateEnum.AuthenticationParam(reservationId, recipientNo, authenticationNumber);
+    }
 }
