+++ xcape-core/src/main/java/com/chadev/xcape/core/exception/ErrorCode.java
@@ -17,6 +17,7 @@ public enum ErrorCode {
     NOT_EXISTENT_PRICE("9007", "존재하지 않는 가격정책입니다."),
     NOT_EXISTENT_RESERVATION_HISTORY("9008", "존재하지 않는 예약기록입니다."),
     NOT_EXISTENT_SCHEDULER("9009", "존재하지 않는 스케줄러입니다."),
+    ALREADY_RESERVATION("9010", "이미 예약된 예약내역입니다."),
 
     // 인증
     AUTHENTICATION_INVALID_PHONE_NUMBER("9100", "예약자의 연락처와 일치하지 않는 연락처입니다."),
