+++ xcape-core/src/main/java/com/chadev/xcape/core/exception/XcapeException.java
@@ -21,4 +21,5 @@ public class XcapeException extends RuntimeException {
     public static XcapeException NOT_EXISTENT_THEME() {return new XcapeException(ErrorCode.NOT_EXISTENT_THEME);}
 
     public static XcapeException NOT_EXISTENT_RESERVATION() {return new XcapeException(ErrorCode.NOT_EXISTENT_RESERVATION);}
+    public static XcapeException ALREADY_RESERVATION() {return new XcapeException(ErrorCode.ALREADY_RESERVATION);}
 }
