+++ xcape-core/src/main/java/com/chadev/xcape/core/service/notification/Header.java
@@ -1,0 +1,12 @@
+package com.chadev.xcape.core.service.notification;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Header {
+    public int resultCode;
+    public String resultMessage;
+    public boolean isSuccessful;
+}
