+++ xcape-core/src/main/java/com/chadev/xcape/core/service/notification/NotificationInterface.java
@@ -1,0 +1,6 @@
+package com.chadev.xcape.core.service.notification;
+
+public interface NotificationInterface<NotificationRequest, NotificationResponse> {
+    NotificationResponse sendMessage(NotificationRequest request);
+
+}
