+++ xcape-core/src/main/java/com/chadev/xcape/core/service/notification/NotificationResponse.java
@@ -1,0 +1,5 @@
+package com.chadev.xcape.core.service.notification;
+
+public interface NotificationResponse {
+    Header getHeader();
+}
